#229 Move session components from admin to trade, add common category for session views
