+++ ui/webui-trade/src/main/java/org/marketcetera/web/session/view/SessionContentCategory.java
@@ -1,0 +1,64 @@
+package org.marketcetera.web.session.view;
+
+import org.marketcetera.web.view.MenuContent;
+
+import com.vaadin.server.Resource;
+import com.vaadin.ui.MenuBar.Command;
+
+/* $License$ */
+
+/**
+ * Provides the menu category for session actions.
+ *
+ * @author <a href="mailto:colin@marketcetera.com">Colin DuPlantis</a>
+ * @version $Id$
+ * @since $Release$
+ */
+public class SessionContentCategory
+        implements MenuContent
+{
+    /* (non-Javadoc)
+     * @see com.marketcetera.web.view.MenuContent#getCaption()
+     */
+    @Override
+    public String getMenuCaption()
+    {
+        return "Sessions";
+    }
+    /* (non-Javadoc)
+     * @see com.marketcetera.web.view.MenuContent#getWeight()
+     */
+    @Override
+    public int getWeight()
+    {
+        return 350;
+    }
+    /* (non-Javadoc)
+     * @see com.marketcetera.web.view.MenuContent#getCategory()
+     */
+    @Override
+    public MenuContent getCategory()
+    {
+        return null;
+    }
+    /* (non-Javadoc)
+     * @see com.marketcetera.web.view.MenuContent#getCommand()
+     */
+    @Override
+    public Command getCommand()
+    {
+        return null;
+    }
+    /* (non-Javadoc)
+     * @see com.marketcetera.web.view.MenuContent#getIcon()
+     */
+    @Override
+    public Resource getMenuIcon()
+    {
+        return null;
+    }
+    /**
+     * sessions menu category
+     */
+    public static final SessionContentCategory instance = new SessionContentCategory();
+}
