+++ ui/webui-trade/src/main/java/org/marketcetera/web/session/view/SessionView.java
@@ -1,4 +1,4 @@
-package org.marketcetera.web.view.sessions;
+package org.marketcetera.web.session.view;
 
 import java.net.Socket;
 import java.util.Collection;
